requirements.txt
cvui==2.7
matplotlib==3.2.2
imutils==0.5.3
opencv_python==4.2.0.34
scipy==1.4.1
numpy==1.18.5
pyserial==3.4
scikit_learn==0.23.2
